fix(ci): stabilize Dubins physics integration on slower CI runners
- Centralize planner RNG seed (11) in tests/integration/conftest.py
- Raise V114 race-duration gate to 180 s (CI ~1.8× slower than dev VM)
- Clean /tmp/fret_physics/dubins_race before each Dubins physics test
