+++ tests/integration/conftest.py
@@ -10,9 +10,26 @@ from __future__ import annotations
 
 from collections.abc import Generator
 
+import numpy as np
 import pytest
 import rclpy
 from rclpy.node import Node
+
+# ARC planners call ``np.random.default_rng()`` without a seed; pin for CI.
+_PLANNER_RNG_SEED = 11
+
+
+@pytest.fixture(autouse=True)
+def _deterministic_planner_rng(monkeypatch: pytest.MonkeyPatch) -> None:
+    """Seed ARC planner RNG so physics integration gates are reproducible."""
+    original_default_rng = np.random.default_rng
+
+    def _seeded_default_rng(seed: int | None = None) -> np.random.Generator:
+        return original_default_rng(
+            _PLANNER_RNG_SEED if seed is None else seed
+        )
+
+    monkeypatch.setattr(np.random, "default_rng", _seeded_default_rng)
 
 
 @pytest.fixture(scope="session")
